Reorganized usecase factories
